Implement car engines
Implement the car engine and basic data structure so we can start
turning on and off car engines.

I have lots to go with the implementation of this.  It doesn't feel
quite right.  Probably the object cloning code and the RuntimeException
extensions, along with no proper error handling.  I think I'll implement
that primarily in the transactions instead.

Signed-off-by: Ryan Y <ryayak1460@protingumas.org>